pyproject.toml
[tool.poetry]
name = "PythonPlaywright"
version = "0.1.0"
description = ""
authors = ["Spencer LaValle"]
readme = "README.md"

[tool.poetry.dependencies]
python = "\u001b[D\u001b[D"
pytest-playwright = "^0.5.2"
ruff = "^0.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
